[8845] fixed some gcc warnings

+++ src/game/Channel.cpp
@@ -455,7 +455,8 @@ void Channel::List(Player* player)
 
             // PLAYER can't see MODERATOR, GAME MASTER, ADMINISTRATOR characters
             // MODERATOR, GAME MASTER, ADMINISTRATOR can see all
-            if (plr && (player->GetSession()->GetSecurity() > SEC_PLAYER || plr->GetSession()->GetSecurity() <= gmLevelInWhoList) &&
+            if (plr && (player->GetSession()->GetSecurity() > SEC_PLAYER ||
+                (uint32)plr->GetSession()->GetSecurity() <= gmLevelInWhoList) &&
                 plr->IsVisibleGloballyFor(player))
             {
                 data << uint64(i->first);
